re-worked signal; fixed tests; added signal compare function in tests
Long algorithm now seems to work without any errors. Values match (rel_tol=1e-7) the expected value at every timestep in the expected result signal (intermediate step results are ignored; but those are computed following the same rules, so this shouldn't be a problem). The only errors in Until tests, now, are the derivatives. These I'm 100% sure are a problem in the source of test data, since a non-zero derivative is immediately followed by a non-equal value.